refactor: split signals into separate file
