+++ packages/pg-gateway/src/auth/password.ts
@@ -1,9 +1,9 @@
 import { createBackendErrorMessage } from '../backend-error.js';
 import type { BufferReader } from '../buffer-reader.js';
 import type { BufferWriter } from '../buffer-writer.js';
-import { closeSignal } from '../connection.js';
 import type { ConnectionState } from '../connection.types';
 import { BackendMessageCode } from '../message-codes';
+import { closeSignal } from '../signals.js';
 import { BaseAuthFlow } from './base-auth-flow';
 
 export type ClearTextPassword = string;
