+++ packages/pg-gateway/src/signals.ts
@@ -1,0 +1,7 @@
+export const tlsUpgradeSignal = Symbol('tls-upgrade');
+export const closeSignal = Symbol('close');
+
+export type TlsUpgradeSignal = typeof tlsUpgradeSignal;
+export type CloseSignal = typeof closeSignal;
+
+export type ConnectionSignal = TlsUpgradeSignal | CloseSignal;
